requirements.txt
asgiref==3.6.0
Django==3.2
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
Pillow==9.4.0
python-decouple==3.8
pytz==2022.7.1
sqlparse==0.4.3
